+++ HISTORY.md
@@ -48,7 +48,10 @@
 * **Translation Updates:**
     * German
 
-### 2022-05-017
+### 2022-05-19
+* blocks, lists, objects, threads: made global Contexts observable 
+
+### 2022-05-17
 * blocks: added experimental private isChangeableTo(type) method
 * blocks, threads: tweaked programmatic blocks-changing